+++ lib/features/recipe/presentation/pages/recipe_view.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:cooking_friend/features/recipe/presentation/widgets/recipe_view_refresh.dart';
 import 'package:cooking_friend/skeleton/constants.dart';
 import 'package:cooking_friend/core/errors/failure.dart';
 import 'package:cooking_friend/features/recipe/business/entities/recipe_entity.dart';
@@ -10,7 +11,6 @@ import 'package:cooking_friend/features/recipe/data/repositories/i_recipe_reposi
 import 'package:cooking_friend/features/recipe/presentation/provider/recipe_getx.dart';
 import 'package:cooking_friend/skeleton/theme/widget/gradient_background.dart';
 import 'package:cooking_friend/global_widget/search_bar_custom.dart';
-import 'package:cooking_friend/global_widget/search_display_card.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
@@ -28,7 +28,7 @@ class _RecipeViewState extends State<RecipeView> {
   final RecipeGetx recipeGetx = Get.find<RecipeGetx>();
   Future<dartz.Either<Failure, List<RecipeEntity>>> recipeToDisplay =
       Completer<dartz.Either<Failure, List<RecipeEntity>>>().future;
-  late final RecipeUseCase recipeService =
+  late final RecipeUseCase recipeUseCase =
       RecipeUseCase(recipeGetx, widget.recipeRepository);
 
   @override
@@ -47,45 +47,6 @@ class _RecipeViewState extends State<RecipeView> {
     });
   }
 
-  void updateFilterRules(String orderBy, String direction) {
-    recipeGetx.currentDirection =
-        OrderByDirection.values.firstWhere((x) => x.paramName == direction);
-
-    recipeGetx.currentOrderBy =
-        RecipeOrderBy.values.firstWhere((x) => x.paramName == orderBy);
-  }
-
-  Future<void> orderBy(List<RecipeEntity> lstRecipe) async {
-    switch (recipeGetx.currentOrderBy) {
-      case RecipeOrderBy.id:
-        if (recipeGetx.currentDirection == OrderByDirection.ascending) {
-          lstRecipe.sort((a, b) => a.idRecipe!.compareTo(b.idRecipe!));
-        } else {
-          lstRecipe.sort((a, b) => b.idRecipe!.compareTo(a.idRecipe!));
-        }
-      case RecipeOrderBy.name:
-        if (recipeGetx.currentDirection == OrderByDirection.ascending) {
-          lstRecipe.sort((a, b) => a.title!.compareTo(b.title!));
-        } else {
-          lstRecipe.sort((a, b) => b.title!.compareTo(a.title!));
-        }
-      case RecipeOrderBy.favorite:
-        if (recipeGetx.currentDirection == OrderByDirection.ascending) {
-          lstRecipe.sort(
-            (a, b) {
-              return b.isFavorite != null && b.isFavorite! == 1 ? 1 : 0;
-            },
-          );
-        } else {
-          lstRecipe.sort(
-            (a, b) {
-              return a.isFavorite != null && a.isFavorite! == 1 ? 1 : 0;
-            },
-          );
-        }
-    }
-  }
-
   @override
   Widget build(BuildContext context) {
     num screenWidth = MediaQuery.of(context).size.width;
@@ -95,7 +56,7 @@ class _RecipeViewState extends State<RecipeView> {
         floatingActionButton: FloatingActionButton(
           onPressed: () async {
             recipeGetx.updateAction(RecipeManagementAction.add);
-            await recipeService.updateList("/recipeAdd", context);
+            await recipeUseCase.updateList("/recipeAdd", context);
           },
           child: const Icon(Icons.add),
         ),
@@ -106,7 +67,7 @@ class _RecipeViewState extends State<RecipeView> {
               child: SearchBarCustom(
                 searchBarController,
                 refreshList,
-                updateFilterRules,
+                recipeUseCase.updateFilterRules,
                 null,
                 RecipeOrderBy.values.map((toElement) {
                   return toElement.paramName;
@@ -125,30 +86,7 @@ class _RecipeViewState extends State<RecipeView> {
                       }, (recipeEntities) {
                         recipeGetx.updateLstRecipeDisplayed(
                             recipeEntities.map((x) => x.toModel()).toList());
-                        return RefreshIndicator(
-                          onRefresh: () => refreshList(),
-                          child: Obx(() {
-                            orderBy(recipeEntities);
-                            return ListView.builder(
-                              physics: const AlwaysScrollableScrollPhysics(),
-                              shrinkWrap: true,
-                              padding: const EdgeInsets.all(8),
-                              itemCount: recipeGetx.lstRecipe.length,
-                              itemBuilder: (BuildContext context, int index) {
-                                String name =
-                                    recipeGetx.lstRecipe[index].title!;
-                                int id = recipeGetx.lstRecipe[index].idRecipe!;
-                                return SearchDisplayCard(
-                                  () => recipeService.clickOnCard(id, context),
-                                  ListTile(
-                                    leading: const Icon(Icons.album),
-                                    title: Text(name),
-                                  ),
-                                );
-                              },
-                            );
-                          }),
-                        );
+                        return RecipeViewRefresh(refreshList(), recipeUseCase, recipeEntities, recipeGetx);
                       });
                       return widgetToDisplay;
                     }
