+++ lib/features/storage/business/entities/storage_modification_entity.dart
@@ -1,5 +1,4 @@
 import 'package:cooking_friend/skeleton/constants.dart';
-
 import 'storage_entity.dart';
 
 class StorageItemModificationEntity {
